Moved the electrs code to pkg/chain/bitcoin
