+++ pkg/chain/bitcoin/electrs_test.go
@@ -1,4 +1,4 @@
-package recovery
+package bitcoin
 
 import (
 	"bytes"
